DERBY-6810: Add regression tests for XXE vulnerability
This change adjusts XMLXXETest.testDerby6807FileAccess() slightly so
that it doesn't add an extra File.separator, which appeared to throw
off the file: scheme parsing on Unix platforms and caused the test to
misbehave.

For me, the test now behaves as expected on both Windows and Linux.

This change also re-adds XMLXXEtest to XMLSuite so that it will get
run by the main Jenkins build scripts and we can see how it behaves
on those platforms.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1686755 13f79535-47bb-0310-9956-ffa450edef68