DERBY-6824: Refactor ShutdownException into shared code area
Having the ShutdownException as a common shared exception class offers no
benefit on its own, but it is an enabler for being able to share the
exception encoding/decoding routines between the network client and the
network server.

There should be no functional changes as a result of this patch, and there
are no new tests added. Subsequent projects intend to build on this change
to introduce functional enhancements and they will include tests for that.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1690384 13f79535-47bb-0310-9956-ffa450edef68