DERBY-6801: Implement MessageUtils class so client/server can share error code.
This patch was contributed by Abhinav Gupta (abhinavgupta2004 at gmail dot com)

This change implements a new class in the "shared" portion of the Derby code
tree. The MessageUtils class contains code which was refactored from various
Engine and Network Server classes, including MessageService and DRDAConnThread.

The purpose of this class is to encapsulate certain details about how error
messages and exceptions are encoded and decoded when passed between the Network
Server and the Network Client, so that the Network Client can decode the
error information to access the arguments to the message when needed.

Nothing about the protocol between the client and server is changed, nor is
anything about the encoding of the data changed. The only change is to make
the logic which implements the encoding/decoding available to both the server
and the client so that they can both use it without duplicating the logic.

As of this change, nothing uses the MessageUtils class. That is, this change
adds the new class, but doesn't remove any of the existing implementations in
the Engine and Network Server code bases.

Follow-on changes will (incrementally) change the Engine, the Network Server,
and the Network Client to use the code in this new MessageUtils class.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1694134 13f79535-47bb-0310-9956-ffa450edef68