DERBY-6802: Change NetworkServer code to use new MessageUtils class DERBY-6823: Refactor uses of DB2_JCC_MAX_EXCEPTION_PARAM_LENGTH
This patch was contributed by Abhinav Gupta (abhinavgupta2004 at gmail dot com)

This change modifies several classes in the Network Server and Engine
code bases to use the new features of the MessageUtils class in the
shared code.

The intent of this change is that no behavior is altered. The same
algorithms, data structures, network protocols, etc. are in use;
the only difference is that the implementation of the error message
and exception encoding logic is now drawn from MessageUtils methods
rather than being spread over several classes in the Network Server
and Engine code.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1694383 13f79535-47bb-0310-9956-ffa450edef68