DERBY-6864: RawDBReaderTest fails on Windows
RawDBReaderTest.runRecoveryScript() opens a FileReader, but doesn't
close it. When the test harness tries to delete the test directory
upon completion of the test, it fails because one of the files in the
directory is still open.

This patch makes runRecoveryScript() open the FileReader in a
try-with-resources statement, so that it is automatically closed when
it goes out of scope.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1728370 13f79535-47bb-0310-9956-ffa450edef68