DERBY-6860: Automatic download of junit.jar broken
Invoke the install_junit target before the setCompilerProperties
target. setCompilerProperties sets the junit property, and the
install_junit raises an error if the junit property is set and
points to a file that doesn't exist. install_junit must therefore
be invoked before the junit property is set in order to be able to
download junit.jar.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1728431 13f79535-47bb-0310-9956-ffa450edef68