DERBY-6828: Network Server fails to start in Czech locale.
This change modifies the drda message properties filename so that it
works more elegantly with the LocalizedResource utility class, and with
the underlying JDK ResourceBundle class on which it is based.

If your locale is cs_CZ, ResourceBundle will look for the message in
messages_cs_CZ.properties. If it is not found there, it will try
with a less specific locale and look for it in messages_cs.properties.

If it is not found there either, it will try with an even less specific
locale and look for it in messages.properties. If we move the English
messages from messages_en.properties to messages.properties, they will
become fallback for messages that lack translation to the requested locale.

The tools messages are already structured like this. There is a
toolsmessages.properties, but not a toolsmessages_en.properties.

The engine messages don't go through LocalizedResource; they have a
different mechanism for falling back to English (using MessageService).

It would be best if we had a single approach for Derby message property
files. However, messages processed by MessageService use a different
naming convention than messages processed by LocalizedResource, so for
the time being it seems the best approach is to rename the drda messages
to follow the LocalizedResource naming convention.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1730839 13f79535-47bb-0310-9956-ffa450edef68