DERBY-1773: cursor updates fail with syntax error when column has an alias
This change enhances the runtime column analysis code so that an updatable
cursor can make a more nuanced decision about whether a column update is
or is not allowed.

Specifically, certain columns may not be updated, if they have been aliased.

Prior to this change, a confusing syntax error message would be delivered
when attempting to update an aliased column. Now, a more clear error message
is delivered, pointing at the fact that the aliased column is not in the
FOR UPDATE list of the cursor.

So the net result is (at least, should be) that the same set of queries are
accepted, but those that are not accepted have a slightly more clear message
issued when they are detected.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1734744 13f79535-47bb-0310-9956-ffa450edef68