DERBY-6881: Test failures with JDK 9-ea b111
ClassLoaderTestSetup casts the context class loader to URLClassLoader
in order to create a new URLClassLoader with the same URLs as the
original class loader. In JDK 9, the context class loader is no longer
a URLClassLoader after JEP 261. This code now results in a
ClassCastException.

The point of ClassLoaderTestSetup is to change the context class
loader to something which is not the system class loader. It does not
have to be a clone of the original context class loader. This patch
therefore changes ClassLoaderTestSetup so that it sets the context
class loader to an empty URLClassLoader which wraps the system class
loader. This avoids the unreliable cast and the resulting
ClassCastException.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1737188 13f79535-47bb-0310-9956-ffa450edef68