DERBY-6881: Test failures with JDK 9-ea b111
Some stack traces are formatted differently in JDK 9. For example,
what used to look like this
  at java.io.FileInputStream.open0(Native Method)
now looks like this
  at java.io.FileInputStream.open0(java.base@9-ea/Native Method)

The old test harness filters out stack traces to get stable results in
the canon-based tests, but it doesn't recognize the new look of the
stack traces.

This patch updates the regular expression that filters out native
methods from stack traces, so that it also recognizes the new format.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1737189 13f79535-47bb-0310-9956-ffa450edef68