DERBY-6885: Remove ReuseFactory
The methods getInteger(), getShort(), getByte(), getLong() and
getBoolean() are now simple wrappers around Integer.valueOf(),
Short.valueOf(), and so on. They no longer have any purpose, and can
be removed. Callers can either call the corresponding valueOf() method
directly or rely on auto-boxing (which use valueOf() implicitly).

The remaining method, getZeroLenByteArray(), which returns a static
instance of an empty byte array, is replaced with a static final field
in the ArrayUtil class.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1740842 13f79535-47bb-0310-9956-ffa450edef68