DERBY-6880: Update failing with java.sql.SQLDataException
This change reverts part of the changes made by revision 1628596 for DERBY-6742.

Specifically, the section of code added to UpdateResultSet.collectAffectedRows
is removed. That code caused problems with certain SQL UPDATE statements
which were compiled with Statement.RETURN_GENERATED_KEYS. The new test cases
added by this change include several examples of such SQL statements.

The JDBC documentation for the intended behavior of UPDATE statements with the
RETURN_GENERATED_KEYS option is unclear; the intended behavior is much
clearer with INSERT statements. Given that I don't understand the intended
behavior, it seems safer to me to return Derby to the previous state for
UPDATE statements; namely, that no attempt is made to compute the set of
generated keys for an UPDATE statement.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1741380 13f79535-47bb-0310-9956-ffa450edef68