DERBY-6854: Make it possible to run Derby tests on early access versions of JDK 9
JDK 9 will move the JDBC classes out of the boot class loader and into
the platform class loader (see JDK-8154189 in the OpenJDK bug
tracker).

Derby has some test and debug code that assumes that the JDBC classes
live in the boot class loader.

1) Some tests create a new class loader in order to make the test load
the Derby classes afresh (SysinfoLocaleTest), or to test against a
different version of Derby (the upgrade tests). To achieve this, they
create a URLClassLoader which contains the Derby jar files they want
to test against, and set the parent of the URLClassLoader to null,
which means that the boot class loader is the parent.

Now that the JDBC classes are not on the boot class loader, this
URLClassLoader is not able to find the JDBC classes, and the tests
fail with NoClassDefFoundErrors when trying to load them.

This is fixed by setting the parent of the created URLClassLoader to
what java.sql.Connection.class.getClassLoader() returns. On JDK 8 and
earlier it returns null, so nothing changes there. On JDK 9 it returns
the platform class loader, which is capable of loading the JDBC
classes.

2) The org.apache.derby.impl.services.bytecode.d_BCValidate class
contains some debug code (only included in the sane jars) which runs
some sanity checks on the generated byte code. To do this, it accesses
the class loader of all the classes in the method signatures.

When getClassLoader() is called on JDBC classes, a RuntimePermission
is needed now. This was not the case before when the JDBC classes were
in the boot class loader, since Class.getClassLoader() does not check
for permissions if the class loader is null. Now this causes an
AccessControlException in tests that run with a security manager. The
tests actually grant the required permission to derby.jar, but the
call is not wrapped in AccessController.doPrivileged(), so it fails
because the permission is not granted to the test code.

The fix is to wrap the call to getClassLoader() in a doPrivileged()
block. Additionally, because RuntimePermission("getClassLoader") is
not a mandatory permission for derby.jar, we need to check if it
raises a security exception. d_BCValidate only needs the class loader
in order to check if it is the same class loader as d_BCValidate's own
class loader. Since Class.getClassLoader() does not require any
permissions if the class has the same class loader as the calling
class, we can conclude that the class loaders are different if a
security exception is raised.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1744989 13f79535-47bb-0310-9956-ffa450edef68