DERBY-6752: Derby tries to load non-existent class AutoloadedDriver40
This patch was contributed by Danoja Dias (danojadias at gmail dot com)

This patch removes the code that tries to load AutoloadedDriver40. That code
was always failing, because the class was removed from the Derby code by
revision 1494482, but the attempt to load the class was wrapped in an
exception catch block so there were no symptoms of the failed class load
attempt.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1754680 13f79535-47bb-0310-9956-ffa450edef68