DERBY-6902: Value out of range error
This change adds several new test cases to ParameterMappingTest, derived
from the original scenario described in the issue.

These test cases are designed to demonstrate the particular exception
discussed in the issue, as well as to demonstrate several ways to use
a CAST operator so that the query executes as desired.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1769904 13f79535-47bb-0310-9956-ffa450edef68