DERBY-6918: Problem with schema name containing a period
When processing DELETE statements in a database schema including
referential integrity constraints that specify ON DELETE CASCADE,
it is easiest to keep the schema name and table name separate,
rather than pasting them together into a compound name joined with
a period, because pasting them together results in ambiguity about
which period separates the schema name from the table name, and
which period was simply part of the schema name proper.

Adjusted the logic in DeleteNode and DMLModStatementNode accordingly.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1772428 13f79535-47bb-0310-9956-ffa450edef68