DERBY-6903: ALTER TABLE ALTER COLUMN resets CYCLE option DERBY-6904: Enhance ALTER TABLE to allow CYCLE option to be changed DERBY-6905: Enhance SYS.SYSCOLUMNS to report the CYCLE value DERBY-6906: Allow a GENERATED column to explicitly specify NO CYCLE
This patch was contributed by Harshvardhan Gupta (harshvardhan145 at gmail dot com)

The patch adds the CYCLE property as a fundamental property associated
with a column similar to current increment value, current increment step etc.

 *  SYS.SYSCOLUMNS extended to hold the cycle value of identity column.
 *  Extended the grammar to accept "cycle","no cycle" in alter clause
    and changed the underlying column descriptor and sequencing objects.
 *  ALTER clause not involving cycle option (i.e increment and restart)
    won't affect the cycle option.

The code flow for the upgrade handling was modelled after Derby-534
for introducing a new in column SYSTRIGGERS. There are new behaviors
for SYSCOLUMNS in both soft- and hard-upgrade scenarios.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1797283 13f79535-47bb-0310-9956-ffa450edef68